refactor: app 시작 단계 분리
